requirements.txt
google-api-python-client  # YouTube API client
langchain                 # LangChain framework
openai                    # For OpenAI models (optional, if using GPT)
python-dotenv             # For environment variables